Move damage tracking logic to `compositor` in `iced_tiny_skia`
